added layers, added Glad to the project

+++ Povox/src/Povox/Events/MouseEvent.h
@@ -34,8 +34,8 @@ namespace Povox {
 		MouseScrolledEvent(float xOffset, float yOffset)
 			: m_XOffset(xOffset), m_YOffset(yOffset) {}
 
-		inline int GetXOffset() const { return m_XOffset;  }
-		inline int GetYOffset() const { return m_YOffset;  }
+		inline float GetXOffset() const { return m_XOffset;  }
+		inline float GetYOffset() const { return m_YOffset;  }
 
 		std::string ToString() const override
 		{
